requirements.txt
attrs>=18.2.0
fire>=0.1.3
gunicorn>=19.9.0
umuus-utils@git+https://github.com/junmakii/umuus-utils.git#egg=umuus_utils-1.0